requirements.txt
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
psycopg2==2.9.5
redis==4.3.4
celery==5.2.7
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.4
requests==2.28.1